+++ tests/integration/test_ipam.py
@@ -1,4 +1,4 @@
-from pynautobot.models.ipam import IpAddresses
+from pynautobot.core.response import Record
 
 
 def test_ip_address_nat_inside_outside_correct_objects(nb_client):
@@ -8,7 +8,15 @@ def test_ip_address_nat_inside_outside_correct_objects(nb_client):
 
     ip_inside_refresh = nb_client.ipam.ip_addresses.get(address="192.0.2.1/32")
 
-    assert isinstance(ip_outside.nat_inside, IpAddresses)
-    assert isinstance(ip_inside_refresh.nat_outside, IpAddresses)
     assert str(ip_outside.nat_inside) == "192.0.2.1/32"
     assert str(ip_inside_refresh.nat_outside) == "192.0.2.2/32"
+
+
+def test_prefixes_successfully_stringify_tags(nb_client):
+    """Validate prefix will properly stringify the tags attribute and they are Record objects."""
+    tag = nb_client.extras.tags.create(name="production", slug="production")
+    prefix = nb_client.ipam.prefixes.create(prefix="192.0.2.0/24", status="active", tags=[tag.id])
+
+    assert str(prefix) == "192.0.2.0/24"
+    assert prefix.tags
+    assert isinstance(prefix.tags[0], Record)