+++ tests/unit/test_query.py
@@ -6,7 +6,11 @@ from pynautobot.core.query import Request
 
 class RequestTestCase(unittest.TestCase):
     def test_get_count(self):
-        test_obj = Request(http_session=Mock(), base="http://localhost:8001/api/dcim/devices", filters={"q": "abcd"},)
+        test_obj = Request(
+            http_session=Mock(),
+            base="http://localhost:8001/api/dcim/devices",
+            filters={"q": "abcd"},
+        )
         test_obj.http_session.get.return_value.json.return_value = {
             "count": 42,
             "next": "http://localhost:8001/api/dcim/devices?limit=1&offset=1&q=abcd",
@@ -29,7 +33,10 @@ class RequestTestCase(unittest.TestCase):
         )
 
     def test_get_count_no_filters(self):
-        test_obj = Request(http_session=Mock(), base="http://localhost:8001/api/dcim/devices",)
+        test_obj = Request(
+            http_session=Mock(),
+            base="http://localhost:8001/api/dcim/devices",
+        )
         test_obj.http_session.get.return_value.json.return_value = {
             "count": 42,
             "next": "http://localhost:8001/api/dcim/devices?limit=1&offset=1",